chore(wallet-config): move networks config utils from the main file
