+++ suite-common/wallet-config/src/networksConfig.ts
@@ -1,49 +1,7 @@
 import { DeviceModelInternal } from '@trezor/connect';
 
-import { Networks, NetworkType, Explorer } from './types';
-
-type NetworkTypeExplorerMap = {
-    [key in NetworkType]: Explorer;
-};
-
-export const getExplorerUrls = (
-    baseUrl: string,
-    networkType: NetworkType,
-    solanaDevnet?: boolean,
-): Explorer => {
-    const networkTypeExplorerMap: NetworkTypeExplorerMap = {
-        bitcoin: {
-            tx: `${baseUrl}/tx/`,
-            account: `${baseUrl}/xpub/`,
-            address: `${baseUrl}/address/`,
-        },
-        ethereum: {
-            tx: `${baseUrl}/tx/`,
-            account: `${baseUrl}/address/`,
-            address: `${baseUrl}/address/`,
-            nft: `${baseUrl}/nft/`,
-        },
-        ripple: {
-            tx: `${baseUrl}/tx/`,
-            account: `${baseUrl}/account/`,
-            address: `${baseUrl}/account/`,
-        },
-        solana: {
-            tx: `${baseUrl}/tx/`,
-            account: `${baseUrl}/account/`,
-            address: `${baseUrl}/account/`,
-            queryString: solanaDevnet ? `?cluster=devnet` : '',
-        },
-        cardano: {
-            tx: `${baseUrl}/tx/`,
-            account: `${baseUrl}/address/`,
-            address: `${baseUrl}/address/`,
-            token: `${baseUrl}/asset/`,
-        },
-    };
-
-    return networkTypeExplorerMap[networkType];
-};
+import { Networks } from './types';
+import { getExplorerUrls } from './getExplorerUrls';
 
 export const networks = {
     btc: {
